requirements.txt
fastapi>=0.110
uvicorn>=0.27
numpy>=1.24
pillow>=10.0
torch>=2.0
scipy>=1.10

